backend/requirement.txt
flask
flask-cors
boto3
pyjwt
passlib
